==> 1-cancer_start.Rmd <==
---
title: "1-data_collection"
author: "Albert Chen"
date: "6/16/2020"
output: pdf_document
---

```{r}
#import data
cancer_train <- read.csv("data/raw/data_set_ALL_AML_train.csv", quote = "")
cancer_test <- read.csv("data/raw/data_set_ALL_AML_independent.csv", quote = "")
cancer_class <- read.csv("data/raw/actual.csv")

#seperate data into genes / numerical readouts / catagorical readouts
seperate_cat_num_readouts <- function(raw_df){
  genes <- raw_df[,c(1:2)]
  genes$index <- c(1:nrow(genes))
  
  num <- raw_df[, seq(3, ncol(raw_df), 2)]
  cat <- raw_df[, seq(4, ncol(raw_df), 2)]
  names(cat) <- names(num)
  
  num <- t(num)
  cat <- t(cat)

  return(list(genes, num, cat))
}

combine_data_with_class <- function(data, class, num_cat) {
  #remap ordinal catagorical values to numerical
  
  data$patient <- row.names(data)
  data$patient <- substring(data$patient, 2)
  data$patient <- sapply(data$patient, as.numeric)
  data <- data[order(data$patient),]
  
  #cat_cat keeps A, M, P if needed for future
  if (num_cat != "cat_cat") {
    data[data == "A"] <- -1 #A - absent
    data[data == "M"] <- 0 #M - marginal
    data[data == "P"] <- 1 #P - present
    data <- lapply(data, function(x) as.numeric(as.character(x)))
    data <- as.data.frame(data)
  }

  combined <- merge(class, data, by="patient")
  
  return(combined)
}

extract_test_or_train_data <- function(data, class, type) {
  #extract test dataset
  cancer_exp <- seperate_cat_num_readouts(data)
  gene <- as.data.frame(cancer_exp[1])
  num <- as.data.frame(cancer_exp[2])
  cat <- as.data.frame(cancer_exp[3])
  
  #update class to numerical
  class$class[class$cancer == "AML"] <- 0
  class$class[class$cancer == "ALL"] <- 1
  
  #combine class with num or cat data
  num_comb <- combine_data_with_class(num, class, "num")
  cat_comb <- combine_data_with_class(cat, class, "cat")
  cat_cat_comb <- combine_data_with_class(cat, class, "cat_cat")
  
  write.csv(gene,"data/interim/gene.csv")
  write.csv(num_comb, paste("data/interim/num_", type, ".csv", sep=''))
  write.csv(cat_comb, paste("data/interim/cat_", type, ".csv", sep=''))
  write.csv(cat_cat_comb, paste("data/interim/cat_cat_", type, ".csv", sep=''))
  
  return(list(gene, num_comb, cat_comb))
}

test_output <- extract_test_or_train_data(cancer_test, cancer_class, "test")
train_output <- extract_test_or_train_data(cancer_train, cancer_class, "train")

gene <- as.data.frame(train_output[1])
num <- as.data.frame(train_output[2])
cat <- as.data.frame(train_output[3])
```
